refactor: update payment service
